#ID014 [BE] - User SubCategory

+++ app/Http/Controllers/SubCategoryController.php
@@ -25,17 +25,24 @@ class SubCategoryController extends Controller
      */
     public function store(StoreSubCategoryRequest $request)
     {
-        $district = SubCategory::create($request->all());
-        return response()->json($district, 201);
+        $subCategory = SubCategory::create($request->all());
+        return response()->json($subCategory, 201);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(SubCategory $subCategory)
+    public function show($subCategory_id)
     {
         try
         {
+            $subCategory = SubCategory::where('id', $subCategory_id)
+            ->first();
+
+            if (!$subCategory) {
+                return response()->json(['message' => 'SubCategory not found'], 404);
+            }
+
             return response()->json($subCategory, 200);
         }
         catch (Exception $exception)
@@ -47,25 +54,48 @@ class SubCategoryController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateSubCategoryRequest $request, SubCategory $subCategory)
+    public function update(UpdateSubCategoryRequest $request, $subCategory_id)
     {
         try
         {
+            $subCategory = SubCategory::where('id', $subCategory_id)
+            ->first();
+
+            if (!$subCategory) {
+                return response()->json(['message' => 'SubCategory not found'], 404);
+            }
+
             $subCategory->update($request->all());
+            
             return response()->json($subCategory, 200);
         }
         catch (Exception $exception)
         {
-            return response()->json(['error' => $exception], 500);
+            return response()->json(['error' => $exception->getMessage()], 500);
         }
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(SubCategory $subCategory)
+    public function destroy($subCategory_id)
     {
+        try
+        {
+            $subCategory = SubCategory::where('id', $subCategory_id)
+            ->first();
+
+            if (!$subCategory) {
+                return response()->json(['message' => 'SubCategory not found'], 404);
+            }
+
             $subCategory->delete();
-            return response()->json(['message' => 'Deleted'], 205);
+
+            return response()->json(['message' => 'Deleted'], 204);
+        }
+        catch (Exception $exception)
+        {
+            return response()->json(['error' => $exception->getMessage()], 500);
+        }
     }
 }
